TWO CHANGES
For one, I added a setaddr method to the class, to override the
underlying handler, and turn the dict-like object into a cdict as well
(class dict) so that values can be treated as key => value or as
property => attribute value.
Secondly, because python version 3 does not like hashtags (#) as comment
charectors in INI files, I modified the default copyright to have
semi-colons instead, which is the new standard (both used to work, now
only semi-colons work).

Secondly, I adjusted the copyright to have a dynamic year (it used to be
staticaly-coded / hard-coded).  Now it will place the current year in a
new copyright at the top of an INI file it creates for you,
automatcailly, the same as if you were to run our copyright tool.