NEW PROGRAM -> getip
This new program obtains the WAN ip of a server.  There is no pretty way
of doing this, so I created one.  This program is bountiful in features.
One of my favorites is that any user can open it, and as long as they
follow the general methodology of previous functions, they can create
their own IP Address obtaining function.  All they have to do is add a
decorator above it,