TWO NEW PROGRAMS (exit code programs)
First program: return-exit-code
-------------------------------

return-exit-code does just that, you call it, and pass in an exit code.
This is a debugging tool for pass thru programs.

Second Program: test-exit-code
------------------------------

test-exit-code tests an exit code by capturing and returning it to you
by printing it to the screen.  After ensuring the program did not have
read permissions of it's own INI file (/home/jlee/etc/wget.ini); I ran:

	(jlee@neptune (/home/jlee/)--> test-exit-code wget-diff
	Error : [13] when executing command:

					 'wget-diff'

	Exit code was the integer(13)

As you can see, test-exit-code told me the correct Exit Code [Error 13]
is the exit code for a permission error, which is what the program
encountered.