ODD: I DO NOT UNDERSTAND THIS
I have version 1.2 in github, on my laptop, everywhere.  But when I INSTALL
pytis with "pip3 install pytis --user" (locally) I find that at some point in
time, I've crated a version 2.  It apparently wasn't commited, but it was built
into the installer thankfully.  Today, I copied it overtop of
src/pytis/getip.py so that I now have version 2.0.  This is what is in this
commit.