Split off and created alias
I finally updated project-bookmarks to track up to 20, as I have been wanting
to do. Additionally,...
I split out project-bookmarks from common, so it could more easily be shared
with others.  Additionally, I copied it to bookmarks, and renamed the aliases
to create a easier to understand set of command names.